Remove .pyc files and unnecessary audio files
The Python bytecode files are unneeded, as they are automatically
generated each time the program is run. As such, they have been
removed from the repository and are no longer being tracked in order to
cut down on space.

Some older, unused audio files are also removed for this same reason.