Fix Settings transition glitch
In the previous build, input would no longer be accepted after exiting
and then re-entering the Settings State. It turns out the cause of this
was "format_input_text()" within KeyBindingList, which was meant to
rewrite the string name for each input button for drawing... but also
ended up changing the input names within the players' key binding
dicts.

Specifically, when KeyBindingList uses "load_bindings()", it passes
the formatted input names to each KeyBinding object as well. These
names are stored in the "input_name" attribute of the KeyBinding,
which is then used to write the new KeyBinding dicts when
"get_bindings()" from KeyBindingList is called. The dicts are then
passed as the actual key binding dicts that the game uses, but since
the input names don't match the required names in Game States'
"get_player_input()" methods (e.g. "Up" would cause
"if input_name == 'up' to return False"), the controls no longer work!

To fix this, "format_input_text()" was moved to the KeyBinding class.
Now KeyBindingList passes the input name unmodified to each KeyBinding,
allowing them to store the proper name for technical purposes. To keep
the names in the correct format graphically, "format_input_text()" is
only used when passing the input name to a KeyBinding's "input_text"
attribute.