Create new Dia UML diagram for revised TitleState
This new diagram outlines all of the changes that will be made in this
branch.

To start off with are some changes to TitleState and moving some of its
functionality over to the new IntroAnimator and OptionList classes.
(Having one class do everything is so totally wrong, especially in
an OO-project like this one!)