UML: Overhaul OptionList's text animations
It has been decided that each OptionList will have a slide-in and
slide-out animation for switching between OptionLists.
To accomplish this, update_text() will be used to decide which of the
animations will be run, since update_text() is called during each
update cycle and the desired animation would need to be updated just as
often.

To keep things neat and tidy, the original method body of update_text()
will be separated into its own method, flash_text().

Of course, the animations will also require additional variables, which
include:
* animation: For deciding on which animation should be shown.
* TEXT_SLIDE_SPEED: For the constant speed value of sliding options in
    and out.
* ListAnimation: An enum that will be used to set the value of
    animation.