OptionList: Automatically slide in when offscreen
OptionLists are smart: they know that they can't be of much use if the
players can't see them. Because of this, they will automatically set
their animation to 'show all' through update() if this is the case.

TitleState will prevent an infinite loop of sliding out and sliding in,
by immediately switching to the next OptionList before calling
update(). That way, the OptionList will be ready to slide in the next
time it is needed.