CharacterAnimation: Add switch_direction()
This is a public method that should be used instead of flip_sprite() by
other classes. The reason for this is that flip_sprite() can still be
called without altering the animation's is_facing_left attribute,
namely when a new animation is loaded via change_animation() and needs
to have its spritesheet flipped to stay in the same direction as the
previous animation.
By contrast, switch_direction() should be used for an actual change in
the character's current direction.