RosterDisplay: Add get_screen_centered_x()
This means that X is no longer a constant, but a value that is
dynamically calculated based on the width of the row Surface.
This adds much more flexibility to the class, as the value of
SLOTS_PER_ROW can be changed at any time and the rest of the class will
automatically reaccommodate it.