Add extra measures if no characters are loaded
It's possible that no characters could be loaded into the game, either
due to read errors or because the players left the character list file
empty.

In such a situation, the roster will have all-blank slots and cannot be
scrolled. The CharacterPreviews will also be set to None, which will
notify CharacterSelectState not to update or draw the previews while
executing update_state() and draw_state().