Add a stack for currently-active Game States
It's possible that two different States be updated and drawn to the
screen at the same, most notably whenever the Settings State is called.

In order for this to be possible in the new game loop algorithm, it is
required that the Game State Manager keep track of all of the States
that are currently in use, as well as have the ability to push off the
top-most State when needed in a last-in first-out style.