Remove State loading and resetting
The addition of the State stack means that there is no longer an
"in-between" phase for a State where it has been initialized but not
loaded, or loaded but needing to be reset.

When a State is no longer needed, it is popped off the stack and
deleted from memory entirely. Any resetting that needs to take place
will happen in the State's init() method when it is created again.

This will hopefully improve performance in the future, now that the
game no longer needs to allocate memory for States that aren't even
being used.