UML: Plan the StateFader class/module
At least four different Game States (Title, Select Character,
Select Stage, Battle) will need a fade-in or fade-out to black
transition when changing States.

To accomplish this, the StateFader class can be added as a component to
any State and will handle all of the fading, as well as switching
States after a fade-out.