Modify intro in case no characters are loaded
If no characters could be loaded, the Intro shouldn't try to draw,
update, or move the CharacterPreviews (since they're set to None and
doing any of those would cause a run-time error).

Also in this situation, the 'No Characters' text will be passed to the
IntroTransition in place of the VS text so that the former will be
wiped-in instead.