Initialize new Game States in a separate Thread
Something I noticed about the new update algorithm is that it causes
the game to lag when opening the Settings State for the first time.
This is probably because it takes a short while to initialize the
entire Settings State, which causes the game to hang because it
uses up the main game loop.

To remedy this, new States are initialized in a separate thread tied
to prepare_next_state(). This method will run concurrently to the main
thread, and once it is finished, the newly-prepared Game State is
stored in the next_state attribute. A simple if-statement in run_game()
ensures that next_state is run as quickly as possible.

The results seem to be a success, as the game no longer hangs when
opening the Settings State. Still, it remains to be seen if this will
remain true when switching to other Game States, so future enhancements
are a possibility.