Move pop_top_state() to top of Stack Management
This is just so that it isn't in-between prepare_next_state() and
run_next_state(), both of which should be beside each other for a more
logical ordering of the methods.

Also, change_state() now calls pop_top_state() internally to make use
of the abstraction that the latter provides.