StateFader: Use a time-based fade speed
Instead of incrementing or decrementing the State Surface alpha by a
fixed value in FADE_RATE, the fade could be made much smoother by
multiplying a FADE_SPEED value by the amount of time elapsed and adding
or subtracting that from the Surface alpha.

This is because the game's frame rate is slightly inconsistent.