Handle State change through OutroTransition
The ID of the next State is passed through the play() method.
Once the animation is finished, change_state(), which is passed
to the class via init(), is called passing the next_state attribute.

This is a logical change because an outro always precedes a State
change; ergo, the outro itself should be responsible for deciding when
the State change happens.