Remove the filepath attribute from Graphic
The Graphic's file path is not used regularly enough to justify it
becoming an attribute in all instances of Graphic and its subclasses.

The one place that did use file paths - the cursor on the Character
Select Screen - has been changed to rely on a new attribute instead:
current_player, which directly stores either 1 or 2 to represent the
appropriate player.