Fix reading of newlines in character_list.txt
Upon testing the game in Linux, it appeared that none of the characters
would load.

The culprit turned out to be the DOS-style line endings in
character_list.txt. In order to make them work on Linux (and
presumably across all other platforms), character_list.txt, as well as
stage_list.txt later down the line, should be read in the Universal
Newline Mode when being opened by Python.