Make a few fixes to align_scroll_arrows()
Since scroll_up_arrow.png is a spritesheet with multiple frames, only
the width of one frame should be passed to calculate_center_position().

When positioning scroll_up_arrow relative to the top of the first
StageThumbnail, the height of the arrow itself must also be subtracted,
as it counts as part of the "distance" between it and the
StageThumbnail's top-left corner.

Finally, using a hard index (like "2" in the previous version) for the
last Thumbnail will not work if NUM_OF_THUMBS were to change.