Change animation sequence on exit of StageSelect
When switching to another State from StageSelectState, reverse the
order in which Graphics are moved. Instead of moving the
BackgroundLines first, move the StagePreview and text first instead.

Remove the position threshold in slide_graphics() in order to allow
Graphics to slide out of the screen during the outro animation
described above.