Pass CharacterSelectState directly to Transitions
Since the Transitions modify the CharacterSelectState's attributes
directly, it's less of a hassle -- and much less verbose -- to just
store a direct reference to the CharacterSelectState as attributes
within the Transitions themselves. This cuts down on all of the
repetitive reference attributes in those Transitions as well.

Giving the Transitions complete access to the State's members will
also prove useful in a future commit.