Remember selections when returning to CharSelect
Specifically, from the StageSelectState. If so, the music is prevented
from restarting and the RosterDisplay and CharacterPreviews are made to
display the characters chosen previously.

This is accomplished by checking whether character_one and
character_two in the StatePass have non-None values. Hence why those
values are cleared when returning to the Title Screen.