Migrate Title State constants to separate file
This will mark the beginning of a new design decision: to have all
constants separated into their module-specific files in the customize
directory.

Not only will this unclutter many of the modules, but it will also
allow both developers and users to easily modify many aspects of the
program in one dedicated location. This is certainly important for a
*customizable* fighting game.